Added structures and Empty classes for Features finished default importer and exporter
- Updated the Importer file to have required elements included already
- Updated the Exporter file to have the required element included already
- Added the empty files for the feature classes
- Separated the Editor into its components
- Updated added the CSS file for the editor
