+++ Pages/Editor/html/Editor.html
@@ -31,18 +31,19 @@
             href="/node_modules/leaflet-toolbar/dist/leaflet.toolbar.css"
         />
         <!-- End of required for leaflet-toolbar -->
-        <!-- The state handler contains states for the program-->
-        <!-- Elliot : 7/6/2022 -->
-        <script id="StateHandler_js" src="/Util/js/StateHandle.js"></script>
 
-        <!-- the callback functions used to change them should be located strictly in this script unless not possible-->
-        <!-- if they are located somewhere else the exact location and reason must be included as a comment within the tags-->
-        <!-- Elliot : 7/8/2022-->
+        <!-- Callbacks for buttons -->
+        <script id="callbacks_js" src="/Pages/Editor/js/Callbacks.js"></script>
+        <!-- End Callbacks for buttons -->
+
+        <!-- Start Editor stylesheet -->
+        <link rel="stylesheet" href="/Pages/Editor/stylesheet/Editor.css" />
+        <!-- End editor stylesheet -->
     </head>
     <body>
         <!--The map that will be export is not directly edited but rather when a element is changed to the map a callback function adds or removes the element to a key:value dictonary
             which is then saved as the html with the embbeded map as js -->
-        <div id="map" >
+        <div id="map">
             <script
                 id="Editor_js"
                 src="/Pages/Editor/js/Editor.js"
