+++ Pages/Editor/js/Exporter.js
@@ -1,6 +1,12 @@
-import * as EZAS from "./Requirements";
+"use strict";
+import * as EZAS from "../js/Requirements.js";
+let gmap;
 
+//listens for an event that tells it to export the map given by the event
 window.addEventListener("exportMap", (e) => {
-    let map = e.detail;
-    map.toString();
+    gmap = e.map_object;
+    exportMap();
 });
+
+//exports the map
+function exportMap() {}
