+++ Pages/Editor/js/Features/CircleFeature.js
@@ -1,0 +1,3 @@
+"use strict";
+import * as EditorRequirements from "./EditorRequirements.js";
+import { GUID } from "./Requirements.js";
