+++ Pages/Editor/js/Features/RectangleFeature.js
@@ -1,0 +1,7 @@
+"use strict";
+import * as EditorRequirements from "./EditorRequirements.js";
+import { GUID } from "./Requirements.js";
+class RectangleFeature extends L.Rectangle {}
+
+export { RectangleFeature };
+export default RectangleFeature;
