+++ Pages/Editor/js/Requirements.js
@@ -1,0 +1,18 @@
+"use strict";
+import GUID from "../js/Util/GUID.js";
+import MapFeature from "../js/Features/MapFeature.js";
+import MarkerFeature from "../js/Features/MarkerFeature.js";
+import PolylineFeature from "../js/Features/PolylineFeature.js";
+import PolygonFeature from "../js/Features/PolygonFeature.js";
+import CircleFeature from "../js/Features/CircleFeature.js";
+import RectangleFeature from "../js/Features/RectangleFeature.js";
+
+export {
+    GUID,
+    MapFeature,
+    MarkerFeature,
+    PolylineFeature,
+    PolygonFeature,
+    CircleFeature,
+    RectangleFeature,
+};
